Move post-carousel component into `post_carousel/`
