+++ src/app/_components/post_carousel/index.tsx
