+++ src/app/_components/post_carousel/style.css
