+++ src/app/page.tsx
@@ -1,7 +1,7 @@
 import { getAllPosts } from '@/lib/posts';
 import BlogList from '@/app/_components/BlogList';
 import { Suspense } from 'react';
-import PostCarousel from '@/app/_components/post-carousel';
+import PostCarousel from '@/app/_components/post_carousel';
 
 export default function Page() {
   const posts = getAllPosts();
